pyproject.toml
[tool.poetry]
name = "fastapi-business-api-portfolio"
version = "0.1.0"
description = "FastAPI business API portfolio"
authors = ["ruan <ruan@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
pydantic = { extras = ["email"], version = "^1.10.0" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21.0"
requests = "^2.31.0"
httpx = "^0.24.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "C90"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
